Update to top level statements
